+++ CHANGELOG.md
@@ -62,11 +62,10 @@
   * Added support for the Mutual TLS Edge Truststore V2 API.
     * [CreateCASetVersion](https://techdocs.akamai.com/mtls-edge-truststore/reference/post-ca-set-version) - Creates a new CA set version.
     * [CloneCASetVersion](https://techdocs.akamai.com/mtls-edge-truststore/reference/post-clone-ca-set-version) - Creates a clone of a CA set version.
-    * [GetCASetVersions](https://techdocs.akamai.com/mtls-edge-truststore/reference/get-ca-set-versions) - Lists all the available CA sets created under the account.
+    * [ListCASetVersions](https://techdocs.akamai.com/mtls-edge-truststore/reference/get-ca-set-versions) - Lists all the available CA set versions created under the account.
     * [GetCASetVersion](https://techdocs.akamai.com/mtls-edge-truststore/reference/get-ca-set-version) - Fetches details of a CA sets version.
     * [UpdateCASetVersion](https://techdocs.akamai.com/mtls-edge-truststore/reference/put-ca-set-version) - Updates a CA sets version.
     * [GetCASetVersionCertificates](https://techdocs.akamai.com/mtls-edge-truststore/reference/get-ca-set-version-certificates) - Lists certificates details of a CA sets version.
-      
 
 
 
@@ -111,6 +110,14 @@
     * Added [ApplyAccountProtectionControls] in all security controls
     * Added [AccountProtection] in export configuration
     * Added [UsernameToSiem] field in siem settings
+
+* mTLS Truststore
+  * Added support for the Mutual TLS Edge Truststore V2 API.
+    * [ActivateCASetVersion](https://techdocs.akamai.com/mtls-edge-truststore/reference/post-activate-ca-set-version) - Activates CA set version.
+    * [DeactivateCASetVersion](https://techdocs.akamai.com/mtls-edge-truststore/reference/post-deactivate-ca-set-version) - Deactivates CA set version.
+    * [GetCASetVersionActivation](https://techdocs.akamai.com/mtls-edge-truststore/reference/get-deployment-request-details) - Fetches CA set version activation.
+    * [ListCASetVersionActivations](https://techdocs.akamai.com/mtls-edge-truststore/reference/get-activations) - Lists activations for CA set version.
+    * [ListCASetActivations](TODO) - Lists activations for CA set.